Cypress Asset Update

+++ targets/TARGET_Cypress/TARGET_PSOC6/psoc6csp/hal/source/cyhal_not_implemented.c
@@ -145,7 +145,7 @@ bool cyhal_has_connection(uint8_t mux, uint8_t outputIdx)
 cy_rslt_t cyhal_connect_trigger(cyhal_source_t source, cyhal_dest_t dest)
 {
     uint8_t muxIdx = cyhal_dest_to_mux_fake[dest];
-    uint8_t destId = dest - cyhal_mux_dest_index_fake[dest];
+    uint8_t destId = (uint8_t)(dest - cyhal_mux_dest_index_fake[dest]);
     uint8_t sourceCount = cyhal_source_count_per_mux[muxIdx];
 
     if (cyhal_has_connection(muxIdx, destId))
@@ -180,7 +180,7 @@ cy_rslt_t cyhal_connect_trigger(cyhal_source_t source, cyhal_dest_t dest)
             {
                 // This destination can be driven by the output of another mux.
                 uint8_t upstreamMuxIdx = cyhal_dest_to_mux_fake[intraDest];
-                uint8_t intraDestId = intraDest - cyhal_mux_dest_index_fake[intraDest];
+                uint8_t intraDestId = (uint8_t)(intraDest - cyhal_mux_dest_index_fake[intraDest]);
                 uint8_t upstreamMuxSourceCount = cyhal_source_count_per_mux[upstreamMuxIdx];
                 cy_rslt_t result = CYHAL_CONNECT_RSLT_NO_CONNECTION;
